# Conflicts:
#	action-api/src/main/java/com/powsybl/action/json/ActionJsonModule.java
#	action-api/src/test/java/com/powsybl/action/json/JsonActionTest.java
#	cgmes/cgmes-conformity/src/main/java/com/powsybl/cgmes/conformity/CgmesConformity1NetworkCatalog.java
#	cgmes/cgmes-conversion/src/main/java/com/powsybl/cgmes/conversion/Conversion.java
#	cgmes/cgmes-conversion/src/main/java/com/powsybl/cgmes/conversion/elements/EquivalentInjectionConversion.java
#	cgmes/cgmes-conversion/src/main/java/com/powsybl/cgmes/conversion/elements/transformers/TwoWindingsTransformerConversion.java
#	cgmes/cgmes-conversion/src/main/java/com/powsybl/cgmes/conversion/export/EquipmentExport.java
#	cgmes/cgmes-conversion/src/main/java/com/powsybl/cgmes/conversion/export/TopologyExport.java
#	cgmes/cgmes-conversion/src/main/java/com/powsybl/cgmes/conversion/naming/AbstractCgmesAliasNamingStrategy.java
#	cgmes/cgmes-conversion/src/main/java/com/powsybl/cgmes/conversion/naming/SimpleCgmesAliasNamingStrategy.java
#	cgmes/cgmes-conversion/src/test/java/com/powsybl/cgmes/conversion/test/export/ExportXmlCompare.java
#	cgmes/cgmes-extensions/src/main/java/com/powsybl/cgmes/extensions/compatibility/CgmesControlAreasSerDe.java
#	cgmes/cgmes-extensions/src/test/java/com/powsybl/cgmes/extensions/compatibility/CgmesControlAreasSerDeTest.java
#	cgmes/cgmes-gl/src/main/java/com/powsybl/cgmes/gl/LinePositionImporter.java
#	commons/src/main/resources/com/powsybl/commons/reports.properties
#	commons/src/main/resources/com/powsybl/commons/reports_fr.properties
#	contingency/contingency-api/src/test/java/com/powsybl/contingency/BoundaryLineContingencyTest.java
#	contingency/contingency-api/src/test/java/com/powsybl/contingency/tasks/BoundaryLineTrippingTest.java
#	iidm/iidm-api/src/main/java/com/powsybl/iidm/network/Bus.java
#	iidm/iidm-api/src/main/java/com/powsybl/iidm/network/TieLine.java
#	iidm/iidm-api/src/main/java/com/powsybl/iidm/network/ValidationUtil.java
#	iidm/iidm-api/src/main/java/com/powsybl/iidm/network/VoltageLevel.java
#	iidm/iidm-api/src/main/java/com/powsybl/iidm/network/util/BoundaryLineData.java
#	iidm/iidm-api/src/main/java/com/powsybl/iidm/network/util/TieLineUtil.java
#	iidm/iidm-api/src/test/java/com/powsybl/iidm/network/util/BoundaryLineDataTest.java
#	iidm/iidm-api/src/test/java/com/powsybl/iidm/network/util/BoundaryLineTestData.java
#	iidm/iidm-criteria/src/test/java/com/powsybl/iidm/criteria/DanglingLineCriterionTest.java
#	iidm/iidm-impl/src/test/java/com/powsybl/iidm/network/impl/util/SVTest.java
#	iidm/iidm-modification/src/test/java/com/powsybl/iidm/modification/scalable/BoundaryLineScalableTest.java
#	iidm/iidm-modification/src/test/java/com/powsybl/iidm/modification/scalable/ProportionalScalableTest.java
#	iidm/iidm-serde/src/main/java/com/powsybl/iidm/serde/AbstractTreeDataExporter.java
#	iidm/iidm-serde/src/main/java/com/powsybl/iidm/serde/DanglingLineSerDe.java
#	iidm/iidm-serde/src/main/java/com/powsybl/iidm/serde/ShuntSerDe.java
#	iidm/iidm-serde/src/main/java/com/powsybl/iidm/serde/TieLineSerDe.java
#	iidm/iidm-tck/src/test/java/com/powsybl/iidm/network/tck/AbstractBoundaryLineTest.java
#	psse/psse-converter/src/test/java/com/powsybl/psse/converter/PsseFullExportTest.java
#	security-analysis/security-analysis-api/src/main/java/com/powsybl/security/json/limitreduction/LimitReductionDeserializer.java

+++ security-analysis/security-analysis-api/src/main/java/com/powsybl/security/json/limitreduction/LimitReductionDeserializer.java
@@ -30,7 +30,7 @@ public class LimitReductionDeserializer extends StdDeserializer<LimitReduction> 
     }
 
     private static final class ParsingContext {
-        float value;
+        double value;
         LimitType limitType;
         Boolean monitoringOnly;
         ContingencyContext contingencyContext;
@@ -43,7 +43,7 @@ public class LimitReductionDeserializer extends StdDeserializer<LimitReduction> 
         ParsingContext context = new ParsingContext();
         JsonUtil.parseObject(parser, fieldName -> parseLimitReduction(parser, deserializationContext, context, fieldName));
         LimitReduction.Builder builder = LimitReduction.builder(checkAttribute(context.limitType, "limitType"),
-                        checkAttribute(context.value, "value"))
+                checkAttribute(context.value, "value"))
                 .withMonitoringOnly(checkAttribute(context.monitoringOnly, "monitoringOnly"))
                 .withContingencyContext(checkAttribute(context.contingencyContext, "contingencyContext"));
         if (context.networkElementCriteria != null) {
@@ -60,7 +60,7 @@ public class LimitReductionDeserializer extends StdDeserializer<LimitReduction> 
         switch (fieldName) {
             case "value" -> {
                 parser.nextToken();
-                context.value = parser.readValueAs(Float.class);
+                context.value = parser.readValueAs(Double.class);
                 return true;
             }
             case "limitType" -> {
